+++ packages/nextra/src/server/compile.ts
@@ -14,6 +14,7 @@ import remarkGfm from 'remark-gfm'
 import remarkMath from 'remark-math'
 import remarkReadingTime from 'remark-reading-time'
 import remarkSmartypants from 'remark-smartypants'
+import { rendererRich, transformerTwoslash } from 'shikiji-twoslash'
 import type { Pluggable, Plugin } from 'unified'
 import type {
   FrontMatter,
@@ -108,12 +109,6 @@ export async function compileMdx(
     rehypePlugins,
     rehypePrettyCodeOptions
   }: MdxOptions = mdxOptions
-
-  if (rehypePrettyCodeOptions) {
-    throw new Error(
-      "`rehypePrettyCodeOptions` is currently unsupported since `rehype-pretty-code` doesn't support `shikiji` package"
-    )
-  }
 
   const format =
     _format === 'detect' ? (filePath.endsWith('.mdx') ? 'mdx' : 'md') : _format
@@ -278,7 +273,23 @@ export async function compileMdx(
         ...(codeHighlight === false
           ? []
           : [
-              [rehypePrettyCode, DEFAULT_REHYPE_PRETTY_CODE_OPTIONS] as any,
+              [
+                rehypePrettyCode,
+                {
+                  ...DEFAULT_REHYPE_PRETTY_CODE_OPTIONS,
+                  // TODO: For some reason I get Error: Cannot find module 'path' in remote content,
+                  // disable twoslash temporarily
+                  transformers: isRemoteContent
+                    ? []
+                    : [
+                        transformerTwoslash({
+                          renderer: rendererRich(),
+                          explicitTrigger: true
+                        })
+                      ],
+                  ...rehypePrettyCodeOptions
+                }
+              ] as any,
               rehypeAttachCodeMeta
             ]),
         [rehypeExtractTocContent, { isRemoteContent }]