+++ packages/nextra/src/server/schemas.ts
@@ -143,7 +143,7 @@ export const menuItemSchema = z.strictObject({
 })
 
 const separatorItemSchema = z.strictObject({
-  title: titleSchema,
+  title: titleSchema.optional(),
   type: z.literal('separator')
 })