+++ packages/nextra/CHANGELOG.md
@@ -1,5 +1,37 @@
 # nextra
 
+## 3.0.0-alpha.17
+
+### Minor Changes
+
+- 30438264: add shikiji twoslash
+
+  Demo feature:
+  https://nextra-v2-na3obnhub-shuding1.vercel.app/docs/guide/twoslash-support
+
+### Patch Changes
+
+- 9f55bd1f: update rehype-pretty-code/shikiji to latest
+
+## 3.0.0-alpha.16
+
+### Minor Changes
+
+- 5a637010: add icons for following languages:
+
+  - GraphQL (`graphql`)
+  - C++ (`c++`, `cpp`)
+  - C# (`csharp`, `c#`, `cs`)
+  - Python (`python`, `py`)
+
+  allow disallow mobile word wrap button in code blocks with `word-wrap=false`
+  meta data setting
+
+### Patch Changes
+
+- 90c129e6: children in Card component is required only if the image prop is
+  true
+
 ## 3.0.0-alpha.15
 
 ### Patch Changes
@@ -192,6 +224,13 @@
 ### Patch Changes
 
 - d8a406b4: add `"sideEffects": false` for better tree-shaking
+
+## 2.13.3
+
+### Patch Changes
+
+- 93b57052: fix `type: 'separator'` zod validation, mark `title` field as
+  optional
 
 ## 2.13.2